Reorganize macros.
